fix the g2o segmentation fault bug

+++ src/camera.cpp
@@ -69,8 +69,8 @@ bool camera::EstimateBoardPose(cv::Mat image)
                     float y = i * _square_size;
                     // Todo: change 3d-coordinate
                     _p3ds.emplace_back(
-                        cv::Point3f(x, y, 0.0));
-                    _p2ds.emplace_back(cv::Point2f(corners[count].x, corners[count].y));
+                        cv::Point3d(x, y, 0.0));
+                    _p2ds.emplace_back(cv::Point2d(corners[count].x, corners[count].y));
                     count++;
                 }
             }
